+++ src/main/java/com/foo/gosucatcher/domain/bucket/exception/UnsupportedBucketException.java
@@ -1,6 +1,7 @@
-package com.foo.gosucatcher.global.error.exception;
+package com.foo.gosucatcher.domain.bucket.exception;
 
 import com.foo.gosucatcher.global.error.ErrorCode;
+import com.foo.gosucatcher.global.error.exception.BusinessException;
 
 public class UnsupportedBucketException extends BusinessException {
