+++ src/main/java/com/foo/gosucatcher/domain/review/domain/Review.java
@@ -17,8 +17,8 @@ import org.hibernate.annotations.Where;
 import com.foo.gosucatcher.domain.expert.domain.Expert;
 import com.foo.gosucatcher.domain.item.domain.SubItem;
 import com.foo.gosucatcher.domain.member.domain.Member;
+import com.foo.gosucatcher.domain.review.exception.UnsupportedReplierException;
 import com.foo.gosucatcher.global.BaseEntity;
-import com.foo.gosucatcher.global.error.exception.UnsupportedUpdateException;
 
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -91,7 +91,7 @@ public class Review extends BaseEntity {
 		long writerId = writer.getId();
 
 		if (writerId != updaterId) {
-			throw new UnsupportedUpdateException(INVALID_UPDATER);
+			throw new UnsupportedReplierException(INVALID_UPDATER);
 		}
 	}
