+++ src/main/java/com/foo/gosucatcher/domain/review/exception/UnsupportedReplierException.java
@@ -1,6 +1,7 @@
-package com.foo.gosucatcher.global.error.exception;
+package com.foo.gosucatcher.domain.review.exception;
 
 import com.foo.gosucatcher.global.error.ErrorCode;
+import com.foo.gosucatcher.global.error.exception.BusinessException;
 
 public class UnsupportedReplierException extends BusinessException {
