--- src/main/java/com/foo/gosucatcher/global/error/exception/InvalidFileSizeException.java
@@ -1,10 +1,0 @@
-package com.foo.gosucatcher.global.error.exception;
-
-import com.foo.gosucatcher.global.error.ErrorCode;
-
-public class InvalidFileSizeException extends BusinessException{
-
-	public InvalidFileSizeException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
