--- src/main/java/com/foo/gosucatcher/global/error/exception/InvalidReplyCountException.java
@@ -1,10 +1,0 @@
-package com.foo.gosucatcher.global.error.exception;
-
-import com.foo.gosucatcher.global.error.ErrorCode;
-
-public class InvalidReplyCountException extends BusinessException{
-
-	public InvalidReplyCountException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
